refactor(tests): reorganise test files by module instead of phase name
Delete test_phase1_unit.py and test_phase2_conversion.py. Redistribute their
tests to files named after what they test:

- test_ha_client_internals.py: add TestGetTemperatureUnit (ha_client.py)
- test_scheduler.py: add TestTemperatureUnitMethods, TestStartupResolveUnit,
  TestCheckUnitChange (scheduler.py)
- test_routes_helpers.py (new): TestToF + TestDeltaToF pure unit tests
- test_routes_coverage.py: add TestTemperatureUnitSettings + route conversion
  integration tests (TestRoomTempConversion, TestScheduleTempConversion,
  TestThermostatTempConversion, TestOverrideTempConversion)

468 tests, 90.28% coverage — threshold maintained.
